refactor(session): extract useMobileCarousel from DesignSession
The mobile carousel's state, refs, scroll handlers and its two sync effects
move whole and contiguous into components/session/useMobileCarousel.ts, so
the cluster's internal hook order and both literal dep arrays are unchanged
and the two effects keep their global position.

compareCollapsed stays in DesignSession: its useState(isMobile) initializer
has to run after useIsMobile() there.
